Protect the route DELETE /api/area/:areaId/reaction/:reactionId of an unauthorized User #129

+++ server/api/area/reaction/reaction.js
@@ -152,9 +152,27 @@ module.exports = function (app, passport, database) {
     passport.authenticate('jwt', { session: false }),
     async (req, res) => {
       try {
+        const area = await database.prisma.AREA.findUnique({
+          where: {
+            id: req.params.areaId
+          },
+          select: {
+            userId: true,
+            Reactions: {
+              select: {
+                id: true
+              }
+            }
+          }
+        })
+        if (!area || area.userId !== req.user.id)
+          return res.status(404).json({ error: 'Area not found' })
+        if (!area.Reactions.find(reaction => reaction.id === req.params.id))
+          return res.status(404).json({ error: 'Reaction not found' })
+
         const reaction = await database.prisma.AREAhasReactions.delete({
           where: {
-            id: Number(req.params.id)
+            id: req.params.id
           }
         })
         res.status(200).json(reaction)
